requirements.txt
asgiref==3.2.10
astroid==2.4.2
bandit==1.6.2
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.7
django-rest-swagger==2.2.0
djangorestframework==3.11.0
drf-yasg==1.17.1
flake8==3.8.3
gitdb==4.0.5
GitPython==3.1.3
idna==2.10
importlib-metadata==1.7.0
inflection==0.5.0
isort==4.3.21
itypes==1.2.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
openapi-codec==1.3.2
packaging==20.4
pbr==5.4.5
psycopg2-binary==2.8.5
pycodestyle==2.6.0
pydocstyle==5.0.2
pyflakes==2.2.0
pylama==7.7.1
pylint==2.5.3
pyparsing==2.4.7
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.0
six==1.15.0
smmap==3.0.4
snowballstemmer==2.0.0
sqlparse==0.3.1
stevedore==2.0.1
toml==0.10.1
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.9
wrapt==1.12.1
zipp==3.1.0
